backend/requirements.txt
asttokens==2.0.5
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
executing==0.8.3
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-core==4.9.2
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
pymongo==4.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
stack-data==0.2.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
